requirements.txt
tensorflow
pandas
numpy
matplotlib
scikit-learn
pillow
streamlit
